Refactor gallery image modal and add image detail view with URL support
